requirements.txt
pandas
APScheduler
python-dotenv
loguru
openpyxl
requests
pyodbc
pymssql
sqlalchemy
pyinstaller
